Fixing type issues
- The GUIDs (Global Unique IDs) are QByteArray instances. In some places
the code was writed to use QString and sometimes writed to use
QByteArray. This works because C++ is doing implicit type conversions.
  I'm removing QString and replacing with QByteArray.

- Changing integer types to the same types used in ninjam protocol

- Replacing raw array in ServerAuthChallengeMessage by QByteArray

+++ Client/src/Common/NinjamController.h
@@ -157,10 +157,10 @@ private:
 
 private slots:
     // ninjam events
-    void on_ninjamServerBpmChanged(short newBpm);
-    void on_ninjamServerBpiChanged(short oldBpi, short newBpi);
-    void on_ninjamAudiointervalCompleted(const Ninjam::User &user, int channelIndex, const QByteArray &encodedAudioData);
-    void on_ninjamAudioIntervalDownloading(const Ninjam::User &user, int channelIndex, int downloadedBytes);
+    void on_ninjamServerBpmChanged(quint16 newBpm);
+    void on_ninjamServerBpiChanged(quint16 oldBpi, quint16 newBpi);
+    void on_ninjamAudiointervalCompleted(const Ninjam::User &user, quint8 channelIndex, const QByteArray &encodedAudioData);
+    void on_ninjamAudioIntervalDownloading(const Ninjam::User &user, quint8 channelIndex, int downloadedBytes);
     void on_ninjamUserChannelCreated(const Ninjam::User &user, const Ninjam::UserChannel &channel);
     void on_ninjamUserChannelRemoved(const Ninjam::User &user, const Ninjam::UserChannel &channel);
     void on_ninjamUserChannelUpdated(const Ninjam::User &user, const Ninjam::UserChannel &channel);
